my_ai_sql_agent/requirements.txt
streamlit
langchain
langchain-community
sqlalchemy
pymysql
mysql-connector-python
